Add user creation command
This commit removes the sample .sqlite file and adds the command
flask create-user.
