+++ .env
@@ -1,7 +1,7 @@
 FLASK_APP=app
 FLASK_ENV=development
-SEC_KEY=
+SEC_KEY=z6Ct_d2Wy0ZcZZVUYD3beI5ZCSsFrR6-f3ZDyn_MW00
 DATABASE=xlsproject.sqlite
-DROPBOX_TOKEN=
-DROPBOX_PATH=
-DROPBOX_INVALID_FILE_PATH=
+DROPBOX_TOKEN=vHsQWZte_5MAAAAAAAACG736FWlk0ZQZLBYb_cIzDt9CQS6RapoQ4F_px8frOgrg
+DROPBOX_PATH="/sheetgo/sheetgo.bmp"
+DROPBOX_INVALID_FILE_PATH="/sheetgo/sample.xlsx"
