+++ app.py
@@ -1,6 +1,7 @@
 from flask import Flask
 from os import environ, path
 from db import db
+from accounts import user_manager
 from api.views import XLSApi, ImageConversionApi, DropboxImageConversionApi
 from settings import BASE_DIR
 
@@ -15,6 +16,7 @@ application.add_url_rule(
     '/api/image/convert/fromdropbox/', view_func=DropboxImageConversionApi.as_view('dropbox_conversion'),
 )
 db.init(application)
+user_manager.init(application)
 
 
 @application.route('/')
